+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 11.1.0
+
+* `Bot`:
+  * Add opportunity to setup proxy
+
 ## 11.0.0
 
 * `Versions`: